make examples (bindings) use iota rebased instead of old iota
