+++ bindings/android/tests/src/main/java/com/etospheres/etopay/examples/GetExchangeRate14.java
@@ -23,20 +23,10 @@ public class GetExchangeRate14 {
             System.out.println("Created and initialized new user.");
 
             // fetch networks from backend
-            String networks = sdk.getNetworks();
+            sdk.getNetworks();
 
-            List<Network> networksList;
-            try {
-                ObjectMapper objectMapper = new ObjectMapper();
-                networksList = objectMapper.readValue(networks, new TypeReference<List<Network>>() {
-                });
-            } catch (JsonProcessingException e) {
-                throw new RuntimeException("Error processing JSON response", e);
-            }
-
-            Network iotaNetwork = networksList.get(0);
             // set the network configuration for the wallet
-            sdk.setNetwork(iotaNetwork.key);
+            sdk.setNetwork("iota_rebased_testnet");
 
             // get exchange rate
             double exchange_rate = sdk.getExchangeRate();
