+++ bindings/android/tests/src/main/java/com/etospheres/etopay/examples/GetWalletTxList19.java
@@ -29,20 +29,10 @@ public class GetWalletTxList19 {
             System.out.println("Created and initialized new wallet.");
 
             // fetch networks from backend
-            String networks = sdk.getNetworks();
+            sdk.getNetworks();
 
-            List<Network> networksList;
-            try {
-                ObjectMapper objectMapper = new ObjectMapper();
-                networksList = objectMapper.readValue(networks, new TypeReference<List<Network>>() {
-                });
-            } catch (JsonProcessingException e) {
-                throw new RuntimeException("Error processing JSON response", e);
-            }
-
-            Network iotaNetwork = networksList.get(0);
             // set the network configuration for the wallet
-            sdk.setNetwork(iotaNetwork.key);
+            sdk.setNetwork("iota_rebased_testnet");
 
             // get wallet_tx_list
             String wallet_tx_list = sdk.getWalletTransactionList(utils.PIN, 0, 10);
