+++ bindings/android/tests/src/main/java/com/etospheres/etopay/examples/SendAmount13.java
@@ -29,20 +29,10 @@ public class SendAmount13 {
             System.out.println("Created and initialized new wallet from mnemonic.");
 
             // fetch networks from backend
-            String networks = sdk.getNetworks();
+            sdk.getNetworks();
 
-            List<Network> networksList;
-            try {
-                ObjectMapper objectMapper = new ObjectMapper();
-                networksList = objectMapper.readValue(networks, new TypeReference<List<Network>>() {
-                });
-            } catch (JsonProcessingException e) {
-                throw new RuntimeException("Error processing JSON response", e);
-            }
-
-            Network iotaNetwork = networksList.get(0);
             // set the network configuration for the wallet
-            sdk.setNetwork(iotaNetwork.key);
+            sdk.setNetwork("iota_rebased_testnet");
 
             // generate receiver address
             String address = sdk.generateNewAddress(utils.PIN);
@@ -51,6 +41,10 @@ public class SendAmount13 {
             // get balance
             double balance = sdk.getWalletBalance(utils.PIN);
             System.out.println("balance: " + balance);
+
+            // estimate gas
+            String estimate = sdk.estimateGas(utils.PIN, address.toString(), 1, "java bindings test".getBytes());
+            System.out.println("estimate: " + estimate);
 
             // send amount
             String tx_id = sdk.sendAmount(utils.PIN, address.toString(), 1, "java bindings test".getBytes());
