+++ bindings/swift/examples/Sources/06_generate_new_address/main.swift
@@ -30,8 +30,8 @@ Task {
         print("migrated wallet from mnemonic")
 
         // Fetch networks from backend
-        let networks = try await sdk.getNetworks()
-        try await sdk.setNetwork(networks[0].key())
+        let _ = try await sdk.getNetworks()
+        try await sdk.setNetwork("iota_rebased_testnet")
         print("retrieved available networks and set the network for the wallet")
 
         // Generate address
