+++ bindings/swift/examples/Sources/14_get_exchange_rate/main.swift
@@ -24,8 +24,8 @@ Task {
         print("initialized new user: \(env.username)")
 
         // Fetch networks from backend
-        let networks = try await sdk.getNetworks()
-        try await sdk.setNetwork(networks[0].key())
+        let _ = try await sdk.getNetworks()
+        try await sdk.setNetwork("iota_rebased_testnet")
         print("retrieved available networks and set the network for the wallet")
 
         let exchange_rate = try await sdk.getExchangeRate()
