+++ bindings/wasm/examples/05-migrate_wallet_from_backup.ts
@@ -26,7 +26,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     // use wallet
     let _address = await sdk.generateNewAddress(PIN);
