+++ bindings/wasm/examples/07-get_balance.ts
@@ -17,7 +17,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     let address = await sdk.generateNewAddress(PIN);
     console.log("Address:", address);
