+++ bindings/wasm/examples/08-create_purchase_request.ts
@@ -18,7 +18,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     let address = await sdk.generateNewAddress(PIN);
     debug(`Generated new IOTA receiver address: ${address}`);
