+++ bindings/wasm/examples/12-change_password.ts
@@ -20,7 +20,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     let _address = await sdk.generateNewAddress(PIN);
 }
