+++ bindings/wasm/examples/14-get_exchange_rate.ts
@@ -11,7 +11,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     let course = await sdk.getExchangeRate();
     console.log(course);
