+++ bindings/wasm/examples/19-get_wallet_tx_list.ts
@@ -17,7 +17,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     let transactions = await sdk.getWalletTransactionList(PIN, 0, 10);  // Get the transaction list
     console.log("Wallet transactions list: " + JSON.stringify(transactions));
