+++ bindings/wasm/examples/20-send_compliment.ts
@@ -30,7 +30,7 @@ async function main() {
     // fetch networks from backend
     let networks = await sdk.getNetworks();
     // set the network configuration for the wallet
-    sdk.setNetwork(networks[0].key);
+    sdk.setNetwork("iota_rebased_testnet");
 
     // Generate new receiver address and fetch current balance
     let address = await sdk.generateNewAddress(PIN);
