+++ sdk/examples/08_create_purchase_request.rs
@@ -27,9 +27,8 @@ async fn main() {
         .unwrap();
 
     // Fetch networks from backend
-    let networks = sdk.get_networks().await.unwrap();
-    let iota_network_key = &networks.first().unwrap().key;
-    sdk.set_network(iota_network_key.to_string()).await.unwrap();
+    let _ = sdk.get_networks().await.unwrap();
+    sdk.set_network("iota_rebased_testnet".to_string()).await.unwrap();
 
     // Generate address
     let address = sdk.generate_new_address(&user.pin).await.unwrap();
