+++ sdk/examples/12_change_password.rs
@@ -28,9 +28,8 @@ async fn main() {
     sdk.set_wallet_password(&user.pin, &new_password).await.unwrap();
 
     // Fetch networks from backend
-    let networks = sdk.get_networks().await.unwrap();
-    let iota_network_key = &networks.first().unwrap().key;
-    sdk.set_network(iota_network_key.to_string()).await.unwrap();
+    let _ = sdk.get_networks().await.unwrap();
+    sdk.set_network("iota_rebased_testnet".to_string()).await.unwrap();
 
     // use wallet
     let _address = sdk.generate_new_address(&user.pin).await.unwrap();
