+++ sdk/examples/13_send_amount.rs
@@ -27,9 +27,8 @@ async fn main() {
         .unwrap();
 
     // Fetch networks from backend
-    let networks = sdk.get_networks().await.unwrap();
-    let iota_network_key = &networks.first().unwrap().key;
-    sdk.set_network(iota_network_key.to_string()).await.unwrap();
+    let _ = sdk.get_networks().await.unwrap();
+    sdk.set_network("iota_rebased_testnet".to_string()).await.unwrap();
 
     // Generate new address
     let recipient_address = sdk.generate_new_address(&user.pin).await.unwrap();
@@ -39,6 +38,7 @@ async fn main() {
     // Send amount
     let amount = dec!(2.0).try_into().unwrap();
     let data = Some("test".to_string().into_bytes());
+
     // estimate gas
     let estimate = sdk
         .estimate_gas(&user.pin, &recipient_address, amount, data.clone())
