+++ sdk/examples/14_get_exchange_rate.rs
@@ -14,9 +14,8 @@ async fn main() {
     sdk.init_user(&user.username).await.unwrap();
 
     // Fetch networks from backend
-    let networks = sdk.get_networks().await.unwrap();
-    let iota_network_key = &networks.first().unwrap().key;
-    sdk.set_network(iota_network_key.to_string()).await.unwrap();
+    let _ = sdk.get_networks().await.unwrap();
+    sdk.set_network("iota_rebased_testnet".to_string()).await.unwrap();
 
     // Get exchange rate
     let exchange_rate = sdk.get_exchange_rate().await.unwrap();
