+++ sdk/examples/19_get_wallet_tx_list.rs
@@ -24,9 +24,8 @@ async fn main() {
     sdk.create_wallet_from_new_mnemonic(&user.pin).await.unwrap();
 
     // Fetch networks from backend
-    let networks = sdk.get_networks().await.unwrap();
-    let iota_network_key = &networks.first().unwrap().key;
-    sdk.set_network(iota_network_key.to_string()).await.unwrap();
+    let _ = sdk.get_networks().await.unwrap();
+    sdk.set_network("iota_rebased_testnet".to_string()).await.unwrap();
 
     // Get wallet tx list
     let wallet_tx_list = sdk.get_wallet_tx_list(&user.pin, 0, 10).await.unwrap();
