+++ sdk/examples/run_all_examples.sh
@@ -18,7 +18,6 @@ examples=(
     "11_reset_pin"
     "12_change_password"
     "13_send_amount"
-    "13_send_amount_iota_rebased"
     "14_get_exchange_rate"
     "16_get_tx_list"
     "18_delete_user"
